Created remaining adapters, DTOs and controllers for entities

+++ src/main/java/com/mastercode/fitmaster/repository/ActivityRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ActivityRepository extends JpaRepository<Activity, Long> {
+public interface ActivityRepository extends JpaRepository<Activity, Integer> {
     Activity getByOrdinalNumber(Integer ordinalNumber);
 }
